+++ test/napkinsketch_book/cookbook_generated_test.clj
@@ -118,9 +118,8 @@
   ts-dates
   (mapv
    (fn*
-    [p1__166058#]
-    (java.time.LocalDate/ofEpochDay
-     (+ 18262 (* (long p1__166058#) 7))))
+    [p1__85921#]
+    (java.time.LocalDate/ofEpochDay (+ 18262 (* (long p1__85921#) 7))))
    (range 52))))
 
 
@@ -132,8 +131,8 @@
    :value
    (mapv
     (fn*
-     [p1__166059#]
-     (+ 100.0 (* 30.0 (Math/sin (* (double p1__166059#) 0.12)))))
+     [p1__85922#]
+     (+ 100.0 (* 30.0 (Math/sin (* (double p1__85922#) 0.12)))))
     (range 52))}))
 
 
@@ -493,7 +492,7 @@
      (and
       (pos? (:points s))
       (some
-       (fn* [p1__166060#] (= "virginica" p1__166060#))
+       (fn* [p1__85923#] (= "virginica" p1__85923#))
        (:texts s)))))
    v67_l323)))
 
@@ -547,8 +546,8 @@
  (->
   {:x (range 20),
    :y
-   (map (fn* [p1__166061#] (Math/sin (/ p1__166061# 3.0))) (range 20)),
-   :change (map (fn* [p1__166062#] (- p1__166062# 10)) (range 20))}
+   (map (fn* [p1__85924#] (Math/sin (/ p1__85924# 3.0))) (range 20)),
+   :change (map (fn* [p1__85925#] (- p1__85925# 10)) (range 20))}
   (sk/view :x :y)
   (sk/lay (sk/point {:color :change}))
   (sk/plot
@@ -669,8 +668,8 @@
    ys
    (mapv
     (fn*
-     [p1__166063#]
-     (+ (* 3 p1__166063#) 5 (* 2 (- (rng/drandom r) 0.5))))
+     [p1__85926#]
+     (+ (* 3 p1__85926#) 5 (* 2 (- (rng/drandom r) 0.5))))
     xs)]
   (->
    {:x xs, :y ys}